requirements.txt
fastapi
uvicorn
streamlit
openai
azure-cognitiveservices-speech
python-dotenv